methods for calculating username and generating password moved from dao layer to service layer
