+++ src/main/java/com/epam/learn/springcore/dao/impl/TraineeDAOImpl.java
@@ -4,7 +4,6 @@ import com.epam.learn.springcore.dao.TraineeDAO;
 import com.epam.learn.springcore.entity.Trainee;
 import com.epam.learn.springcore.storage.TraineeStorage;
 import lombok.extern.log4j.Log4j2;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
@@ -20,12 +19,6 @@ public class TraineeDAOImpl implements TraineeDAO {
 
     @Override
     public void create(Trainee trainee) {
-        log.info("Calculating trainee username");
-        trainee.setUsername(generateUsername(trainee.getFirstName(), trainee.getLastName()));
-        log.info("Trainee username calculated");
-        log.info("Generating trainee password");
-        trainee.setPassword(generateRandomPassword());
-        log.info("Trainee password generated");
         log.info("Saving trainer to storage");
         traineeStorage.getStorage().put(trainee.getUsername(), trainee);
         log.info("Trainer saved to storage");
@@ -65,18 +58,4 @@ public class TraineeDAOImpl implements TraineeDAO {
         }
     }
 
-    private String generateUsername(String firstName, String lastName) {
-        String baseUsername = firstName + "." + lastName;
-        int count = 1;
-        String finalUsername = baseUsername;
-        while (select(finalUsername) != null) {
-            finalUsername = baseUsername + count;
-            count++;
-        }
-        return finalUsername;
-    }
-
-    private String generateRandomPassword() {
-        return RandomStringUtils.secure().next(10);
-    }
 }
