+++ src/main/java/com/epam/learn/springcore/service/TrainerService.java
@@ -3,6 +3,7 @@ package com.epam.learn.springcore.service;
 import com.epam.learn.springcore.dao.TrainerDAO;
 import com.epam.learn.springcore.entity.Trainer;
 import lombok.extern.log4j.Log4j2;
+import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +14,12 @@ public class TrainerService {
     private TrainerDAO trainerDAO;
 
     public void createTrainer(Trainer trainer) {
+        log.info("Calculating trainer username");
+        trainer.setUsername(calculateUsername(trainer.getFirstName(), trainer.getLastName()));
+        log.info("Trainer username calculated");
+        log.info("Generating trainer password");
+        trainer.setPassword(generateRandomPassword());
+        log.info("Trainer password generated");
         log.info("Creating trainer: {}", trainer);
         trainerDAO.create(trainer);
         log.info("Successfully created trainer: {}", trainer);
@@ -32,6 +39,21 @@ public class TrainerService {
     public Trainer selectTrainer(String username) {
         log.info("Selecting trainer: {}", username);
         return trainerDAO.select(username);
+    }
+
+    private String calculateUsername(String firstName, String lastName) {
+        String baseUsername = firstName + "." + lastName;
+        int count = 1;
+        String finalUsername = baseUsername;
+        while (trainerDAO.select(finalUsername) != null) {
+            finalUsername = baseUsername + count;
+            count++;
+        }
+        return finalUsername;
+    }
+
+    private String generateRandomPassword() {
+        return RandomStringUtils.secure().next(10);
     }
 
 }
