+++ src/test/java/com/epam/learn/springcore/StorageAndDAOTests.java
@@ -57,24 +57,13 @@ class StorageAndDAOTests {
         Trainer trainer = Trainer.builder()
                 .firstName("Sven")
                 .lastName("Nielsen")
+                .username("Sven.Nielsen")
                 .isActive(true)
                 .trainingType(TrainingType.STRETCHING)
                 .build();
         trainerDAO.create(trainer);
         assertEquals(trainerStorage.getStorage().size(), 2);
         assertNotNull(trainerDAO.select("Sven.Nielsen"));
-    }
-
-    @Test
-    void testCreateTrainerWithSameUsername() {
-        Trainer trainer = Trainer.builder()
-                .firstName("Sergii")
-                .lastName("Vakaliuk")
-                .isActive(true)
-                .trainingType(TrainingType.STRETCHING)
-                .build();
-        trainerDAO.create(trainer);
-        assertNotNull(trainerDAO.select("Sergii.Vakaliuk1"));
     }
 
     @Test
@@ -95,6 +84,7 @@ class StorageAndDAOTests {
         Trainee trainee = Trainee.builder()
                 .firstName("Charles")
                 .lastName("Benson")
+                .username("Charles.Benson")
                 .isActive(true)
                 .dateOfBirth(new Date())
                 .address("321 Some St Sometown CA 54321 USA")
@@ -102,29 +92,6 @@ class StorageAndDAOTests {
         traineeDAO.create(trainee);
         assertEquals(3, traineeStorage.getStorage().size());
         assertNotNull(traineeDAO.select("Charles.Benson"));
-    }
-
-    @Test
-    void testCreateTraineesWithSameUsername() {
-        Trainee trainee1 = Trainee.builder()
-                .firstName("Mary")
-                .lastName("Public")
-                .isActive(true)
-                .dateOfBirth(new Date())
-                .address("555 Anywhere St Anytown CO 55555 USA")
-                .build();
-        traineeDAO.create(trainee1);
-        assertNotNull(traineeDAO.select("Mary.Public1"));
-        Trainee trainee2 = Trainee.builder()
-                .firstName("Mary")
-                .lastName("Public")
-                .isActive(false)
-                .dateOfBirth(new Date())
-                .address("222 Somewhere St Itown JO 22222 USA")
-                .build();
-        traineeDAO.create(trainee2);
-        assertNotNull(traineeDAO.select("Mary.Public2"));
-        assertEquals(4, traineeStorage.getStorage().size());
     }
 
     @Test
